# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Cognitive Substrate API/BFF
  version: 0.1.0
  description: |
    The API/BFF gateway for the Cognitive Substrate workbench. The gateway
    bridges the front-end UI to the cognitive pipeline by:

      - accepting user messages and publishing them to `experience.raw` Kafka,
      - streaming `interaction.response` events back to subscribed clients
        over Server-Sent Events,
      - serving read-only views over OpenSearch indexes (memories, audit
        traces, agent activity, policy snapshots),
      - exposing collector control endpoints for Aiven-managed services.

    The cognitive pipeline itself runs as a set of workers behind Kafka and
    is not exposed directly. All write traffic enters via
    `POST /api/sessions/{sessionId}/messages`; all read-back traffic returns
    on `/stream` (SSE) or via the read-only routes listed below.

    Roadmap status (see policy.ts for stage gates):
      - Stage 4 routes (`/policy`) read live data from OpenSearch.
      - Stage 6+ routes (`/agents`, `/selfmod`, `/identity`, `/goals`) return
        stub responses until the corresponding backend stages land.
  license:
    name: PolyForm-Noncommercial-1.0.0
    identifier: PolyForm-Noncommercial-1.0.0
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://{host}
    description: Deployed environment
    variables:
      host:
        default: api.cognitive-substrate.local
...
security: []
...
paths:
...
  /api/sessions/{sessionId}/memories:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [memories]
      summary: List session memories
      description: Returns episodic memories scoped to `sessionId`, ordered by recency.
      operationId: listSessionMemories
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 20
          description: Maximum number of memories to return.
      responses:
        "200":
          description: Memory list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoriesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/sessions/{sessionId}/memories/search:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [memories]
      summary: Hybrid-search semantic memories
      description: |
        Runs a hybrid BM25 + k-NN query against `memory_semantic`. Empty
        queries short-circuit to an empty result set.
      operationId: searchMemories
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          description: Free-text query.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 10
        - in: query
          name: mode
          schema:
            type: string
            enum: [quality, efficient, hybrid, legacy]
            default: legacy
          description: |
            Retrieval lane that selects which `knn_vector` field is used.
            See `packages/memory-opensearch/src/query-builder.ts`.
      responses:
        "200":
          description: Search results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoriesResponse"
  /api/sessions/{sessionId}/memories/trace:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [memories]
      summary: Recent audit / trace events for the session
      description: |
        Reads recent records from the `audit_events` index and projects
        them onto the BFF's `TraceEvent` shape so the workbench can
        render the pipeline timeline.
      operationId: listTraceEvents
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 50
      responses:
        "200":
          description: Trace event list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceEventList"
...
components:
  parameters:
    SessionId:
      in: path
      name: sessionId
      required: true
      schema:
        type: string
      description: Session identifier returned by `POST /api/sessions`.
  responses:
    BadRequest:
      description: The request payload was invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
      example:
        error: "sessionId is required"
...
    Memory:
      type: object
      required: [memoryId, index, summary, importanceScore, score]
      properties:
        memoryId:
          type: string
        index:
          type: string
          description: OpenSearch index the memory was sourced from.
        summary:
          type: string
        importanceScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
        score:
          type: number
          format: float
          description: Search-time relevance score.
        tags:
          type: array
          items:
            type: string
        lastRetrieved:
          type: string
          format: date-time
    MemoriesResponse:
      type: object
      required: [memories, total]
      properties:
        memories:
          type: array
          items:
            $ref: "#/components/schemas/Memory"
        total:
          type: integer
          minimum: 0
    TraceStage:
      type: string
      enum: [received, embedding, indexed, reasoning, arbitrating, complete, failed]
    TraceEvent:
      type: object
      required: [eventId, sessionId, timestamp, stage]
      properties:
        eventId:
          type: string
        sessionId:
          type: string
        timestamp:
          type: string
          format: date-time
        stage:
          $ref: "#/components/schemas/TraceStage"
        detail:
          type: string
    TraceEventList:
      type: object
      required: [events, total]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/TraceEvent"
        total:
          type: integer
          minimum: 0
...